# Separate Graph Structure and Node/Edge Labels
The existing API was a graph which held different types of Nodes/Edges. Different implementations of Nodes/Edges would store different data. For example, for a chess game, a BoardTile could store a coordinate and which piece is on this tile (i.e. Queen, C8).
Now, Node and Edge are concrete classes and not meant to be overridden. They are now only used to store information about the graph structure (i.e. which nodes connect to which, and in what direction?) They do not store any domain-specific data.
To store data about nodes and edges, we will now use LabeledGraph instead. This design has some advantages:
    1. Many operations, such as getting a node's neighbors or performing Depth First Search, only depend on graph structure. Now, their classes can be defined more simply in terms of Node and Edge objects, instead of parameterized "N implements Node" and "E implements Edge" types.
    2. LabeledGraph allows us to store multiple layers of information about nodes/edges of a graph on top of a common network. This prevents duplicate data and reduces the risk of data inconsistency. If the original graph's structure is updated, the LabeledGraph will be automatically updated.
    3. Developers do not need to implement Node or Edge, and can instead quickly create a graph and work with its structure without implementing any extra classes.
    4. Makes tests simpler (No more parameterized GraphTest/NodeTest/EdgeTest)